+++ cpp/src/scheduler/Cost.h
@@ -23,7 +23,7 @@ namespace engine {
  * call from scheduler;
  */
 std::vector<uint64_t>
-PickToMove(const TaskTable &task_table, const CacheMgr &cache_mgr, double limit);
+PickToMove(const TaskTable &task_table, const CacheMgr &cache_mgr, uint64_t limit);
 
 
 /*
